==> recommender/1_1_data_cleaning_anime.ipynb <==
# ANIME RECOMMENDER SYSTEM - DATA CLEANING - ANIME

# basic library
import numpy as np 
import pandas as pd 
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
import pickle
warnings.filterwarnings('ignore')

# load data
anime_df = pd.read_csv('dataset/anime.csv')
anime_df.head()

## CLEANING NULL - DUPLICATES

anime_df.info()

anime_df['ranked'].replace({np.nan: 0}, inplace=True)
anime_df['title'].replace({np.nan: '-'}, inplace=True)
anime_df['synopsis'].replace({np.nan: 'Unknown'}, inplace=True)

anime_df.info()

anime_df = anime_df.drop_duplicates()

anime_df.info()

# drop because all unknown
anime_df = anime_df.drop(['season_aired'], axis=1)

## CLEANING DATA TYPES AND INVALID VALUES

anime_df.rename(columns={'anime_rating': 'score'}, inplace=True)
print(sorted(pd.unique(anime_df['score'])))

def to_num(col):
    # replace 'Unknown' to 0
    anime_df[col].replace({'Unknown': 0}, inplace=True)
    anime_df[col].replace({'?': 0}, inplace=True)

    # convert to float
    anime_df[col] = pd.to_numeric(anime_df[col], downcast='float')

col_to_num_list = ['episode_count', 'score', 'rating_count']
for i in col_to_num_list:
    to_num(i)

anime_df.info()

anime_df.head()

# list cleaning (string)
anime_df['genre'] = anime_df['genre'].str.replace('[','')
anime_df['genre'] = anime_df['genre'].str.replace(']','')
anime_df['genre'] = anime_df['genre'].str.replace('\'','')

anime_df.head()

## EXPORT TO CSV

anime_df.to_csv('dataset/cleaned_dataset/anime.csv', index=False)

with open('./dataset/cleaned_dataset/anime.pkl','wb') as f:
    pickle.dump(anime_df, f)
