==> recommender/1_2_data_cleaning_rating.ipynb <==
# ANIME RECOMMENDER SYSTEM - DATA CLEANING - RATING

# basic library
import numpy as np 
import pandas as pd 
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
import pickle
warnings.filterwarnings('ignore')

# load data
rating_df = pd.read_csv('dataset/rating.csv')
rating_df.head()

rating_df.count()

rating_df.info()

## CLEANING DATA TYPES AND INVALID VALUES

print(sorted(pd.unique(rating_df['is_rewatching'])))

print(sorted(pd.unique(rating_df['status'])))

# drop because all unknown and 0
rating_df = rating_df.drop(['is_rewatching'], axis=1)
# drop because all 2
rating_df = rating_df.drop(['status'], axis=1)

rating_df = rating_df.drop(['anime_title'], axis=1)

rating_df.to_csv('dataset/cleaned_dataset/rating.csv', index=False)

with open('./dataset/cleaned_dataset/rating.pkl','wb') as f:
    pickle.dump(rating_df, f)
