==> recommender/3_2_preprocessing_rating.ipynb <==
# ANIME RECOMMENDER SYSTEM - PREPROCESSING - RATING

# basic library
import numpy as np 
import pandas as pd 
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
warnings.filterwarnings('ignore')
import itertools
import collections
import pickle

# load data
rating_df = pd.read_csv('dataset/cleaned_dataset/rating.csv')
rating_df.head()

rating_df.info()

rating_df.count()

rating_df.drop(rating_df.loc[rating_df['score']==0].index, inplace=True)
rating_df.drop(rating_df.loc[rating_df['num_watched_episodes']==0].index, inplace=True)
rating_df.drop(rating_df.loc[rating_df['created_at']==0].index, inplace=True)
rating_df.drop(rating_df.loc[rating_df['updated_at']==0].index, inplace=True)

rating_df.count()

## Preprocessing step:

1. scaling (numerical features)

### scaling
- score
- num_watched_episodes
- created_at
- updated_at

numerical_col = ['score', 'num_watched_episodes', 'created_at', 'updated_at']

from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()

for col in numerical_col:
    rating_df[col] = scaler.fit_transform(rating_df[[col]])
rating_df.head()

# export preprocessed result to df.

rating_df.to_csv('dataset/processed_dataset/rating.csv', index=False)
