==> recommender/4_collaborative_filtering.ipynb <==
# ANIME RECOMMENDER SYSTEM - COLABORATIVE FILTERING

# basic library
import numpy as np 
import pandas as pd 
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
import pickle
warnings.filterwarnings('ignore')

# # load data
# rating_df = pd.read_csv('dataset/processed_dataset/rating.csv')

# load data
file = open('dataset/processed_dataset/pivot_tab.pkl', 'rb')
pivot_tab = pickle.load(file)

# pivot_tab = rating_df.pivot_table(index=['user_id'], columns=['anime_id'], values='rating').fillna(0)
pivot_tab.head()

pivot_tab = pivot_tab.T

from scipy.sparse import csr_matrix
from sklearn.neighbors import NearestNeighbors
from sklearn.metrics.pairwise import cosine_similarity

cosine_sim_pivot = csr_matrix(pivot_tab.values)

item_similarity = cosine_similarity(cosine_sim_pivot)
user_similarity = cosine_similarity(cosine_sim_pivot.T)
item_sim_df = pd.DataFrame(item_similarity, index = pivot_tab.index, columns = pivot_tab.index)
user_sim_df = pd.DataFrame(user_similarity, index = pivot_tab.columns, columns = pivot_tab.columns)

def get_rec_collaborative_anime_id(anime_id):
    rec = item_sim_df.sort_values(by = anime_id, ascending = False).index[1:11]
    return pd.DataFrame(rec)

def get_rec_collaborative_user_id(user_id):
    rec = user_sim_df.sort_values(by = user_id, ascending = False).index[1:11]
    return pd.DataFrame(rec)

21019

get_rec_collaborative_user_id(2)

get_rec_collaborative_anime_id(1)

import pickle

pickle.dump(item_sim_df, open('pickle/item_sim_df.pickle', 'wb'))

pickle.dump(user_sim_df, open('pickle/user_sim_df.pickle', 'wb'))
