==> recommender/4_2_merge_df.ipynb <==
# ANIME RECOMMENDER SYSTEM - MERGE ANIME & RATING DF - PIVOT TABLE - IMG WITHOUT PCA

# basic library
import numpy as np 
import pandas as pd 
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
warnings.filterwarnings('ignore')
import itertools
import collections
import pickle

# load data
anime_df = pd.read_csv('dataset/processed_dataset/anime_1.csv')
anime_df.head(10)

anime_df.info()

# load data
rating_df = pd.read_csv('dataset/processed_dataset/rating.csv')
rating_df.head()

rating_df.info()

create dummy rating

print avg to create dummy rating

# print average
def printAvg(df, col):
    print("Average: " + str(df[col].mean()))

printAvg(rating_df, 'score')
printAvg(rating_df, 'num_watched_episodes')
printAvg(rating_df, 'created_at')
printAvg(rating_df, 'updated_at')

list_user_id = []
list_score = []
list_num_watched_episodes = []
list_created_at = []
list_updated_at = []
list_anime_id = []

for i, row in anime_df.iterrows():
    list_user_id.append(1)
    list_score.append(0.742121)
    list_num_watched_episodes.append(0)
    list_created_at.append(0.364752)
    list_updated_at.append(0.417971)
    list_anime_id.append(row.anime_id)

list_tuples = list(zip(list_user_id, list_score, list_num_watched_episodes, list_created_at, list_updated_at, list_anime_id)) 

rating_dummy = pd.DataFrame(list_tuples, columns=['user_id', 'score', 'num_watched_episodes', 'created_at', 'updated_at', 'anime_id']) 

rating_dummy.head()

rating_df = rating_df.append(rating_dummy)

rating_df.rename(columns = {'score':'user_score'}, inplace = True)

merge_df = pd.merge(rating_df, anime_df, how='left')
merge_df.head()

merge_df.count()

pivot_tab = rating_df.pivot_table(index=['user_id'], columns=['anime_id'], values='user_score').fillna(0)
pivot_tab.head()

merge_df.to_csv('dataset/processed_dataset/merged_1.csv', index=False)

with open('./dataset/processed_dataset/pivot_tab_1.pkl','wb') as f:
    pickle.dump(pivot_tab, f)
