==> recommender/5_content_based_filtering.ipynb <==
# ANIME RECOMMENDER SYSTEM - CONTENT-BASED FILTERING

# basic library
import numpy as np 
import pandas as pd 
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
warnings.filterwarnings('ignore')

# load data
anime_df = pd.read_csv('dataset/processed_dataset/anime.csv')

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import sigmoid_kernel

def get_tf_matrix(col):
    tf = TfidfVectorizer(stop_words='english')
    anime_df[col] = anime_df[col].fillna('')

    tf_matrix = tf.fit_transform(anime_df[col])
    tf_matrix.shape
    return tf_matrix

def get_cosine(tf_matrix):
    # compute cosine similarity matrix
    cosine_sim = sigmoid_kernel(tf_matrix, tf_matrix)
    return cosine_sim

# get indices for every anime
indices = pd.Series(anime_df.index, index=anime_df['title']).drop_duplicates()

# list to string
col_list = ['genre']
for i in col_list:
    anime_df[i] = anime_df[i].astype(str)

# synopsis
tf_matrix_sypnopsis = get_tf_matrix('synopsis')
cosine_sim_sypnopsis = get_cosine(tf_matrix_sypnopsis)

# genre
tf_matrix_genres = get_tf_matrix('genre')
cosine_sim_genres = get_cosine(tf_matrix_genres)

#studio
tf_matrix_studios = get_tf_matrix('studio')
cosine_sim_studios = get_cosine(tf_matrix_studios)

header = ['anime_id', 'title', 'score', 'rating_count', 'ranked', 'popularity', 'members', 'type', 'studio', 'synopsis', 'episode_count', 'genre', 'url', 'img']

# create soup combining feature:
feature = ['genre', 'type', 'studio', 'synopsis']
for i in feature:
    anime_df = anime_df.astype(str)
anime_df['soup'] = anime_df[feature].apply(lambda row: '  '.join(row.values.astype(str)), axis=1)

tf_matrix_soup = get_tf_matrix('soup')
cosine_sim_soup = get_cosine(tf_matrix_soup)

# get recommendation function for content based filtering
def get_rec_content(title, cosine_sim):
    # get index of title
    idx = indices[title]
    
    # pairwise similarity score 
    sim = list(enumerate(cosine_sim[idx]))
    sim = sorted(sim, key=lambda x: x[1], reverse=True)
    sim = sim[1:11]

    # indices sim 
    anime_indices = [i[0] for i in sim]
    return anime_df['title'].iloc[anime_indices]

get_rec_content('Haikyuu!!', cosine_sim_sypnopsis)

get_rec_content('Haikyuu!!', cosine_sim_genres)

get_rec_content('Haikyuu!!', cosine_sim_studios)

get_rec_content('Haikyuu!!', cosine_sim_soup)

import pickle

pickle.dump(cosine_sim_soup, open('pickle/cosine_sim_soup.pickle', 'wb'))

# get indices for every anime
indices_id = pd.Series(anime_df.index, index=anime_df['anime_id']).drop_duplicates()

pickle.dump(indices_id, open('pickle/indices_id.pickle', 'wb'))
