==> recommender/7_feature_selection.ipynb <==
# ANIME RECOMMENDER SYSTEM - FEATURE SELECTION FOR TWO TOWER NEURAL NET

# basic library
import numpy as np 
import pandas as pd 
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
warnings.filterwarnings('ignore')
import itertools
import collections
import pickle

# load data
merged_df = pd.read_csv('dataset/processed_dataset/merged.csv')
merged_df.head()

all column:

user_id,user_score,num_watched_episodes,created_at,updated_at,anime_id,score,rating_count,ranked,popularity,members,type,studio,episode_count,Action,Adult Cast,Adventure,Anthropomorphic,Avant Garde,Award Winning,Boys Love,CGDCT,Childcare,Combat Sports,Comedy,Crossdressing,Delinquents,Detective,Drama,Ecchi,Educational,Erotica,Fantasy,Gag Humor,Girls Love,Gore,Gourmet,Harem,Hentai,High Stakes Game,Historical,Horror,Idols (Female),Idols (Male),Isekai,Iyashikei,Josei,Kids,Love Polygon,Magical Sex Shift,Mahou Shoujo,Martial Arts,Mecha,Medical,Military,Music,Mystery,Mythology,Organized Crime,Otaku Culture,Parody,Performing Arts,Pets,Psychological,Racing,Reincarnation,Reverse Harem,Romance,Romantic Subtext,Samurai,School,Sci-Fi,Seinen,Shoujo,Shounen,Showbiz,Slice of Life,Space,Sports,Strategy Game,Super Power,Supernatural,Survival,Suspense,Team Sports,Time Travel,Vampire,Video Game,Visual Arts,Workplace,0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,0.1,1.1,2.1,3.1,4.1,5.1,6.1,7.1,8.1,9.1,10.1,11.1,12.1,13.1,14.1,15.1,16.1,17.1,18.1,19.1,20.1,21.1,22.1,23.1,24.1,25.1,26.1,27.1,28.1,29.1,30.1,31.1,32.1,33.1,34.1,35.1,36.1,37.1,38.1,39.1,40.1,41.1,42.1,43.1,44.1,45.1,46.1,47.1,48.1,49.1,50.1,51.1,52.1,53.1,54.1,55.1,56.1,57.1,58.1,59.1,60.1,61.1,62.1,63.1,64.1,65.1,66.1,67.1,68.1,69.1,70.1,71.1,72.1,73.1,74.1,75.1,76.1,77.1,78.1,79.1,80.1,81.1,82.1,83.1,84.1,85.1,86.1,87.1,88.1,89.1,90.1,91.1,92.1,93.1,94.1,95.1,96.1,97.1,98.1,99.1

list:
- case 1    : user-> all, item-> no word no img
- case 2    : user-> all, item-> no img
- case 3    : user-> all, item-> all 

feature selection:
delete: ranked, popularity, rating_count

merged_df.info()

all_col = ['user_id', 'user_score', 'num_watched_episodes','created_at','updated_at', 'anime_id', 'score','members', 'type', 'studio', 'episode_count', 
'Action','Adult Cast','Adventure','Anthropomorphic','Avant Garde','Award Winning','Boys Love','CGDCT','Childcare','Combat Sports','Comedy',
'Crossdressing','Delinquents','Detective','Drama','Ecchi','Educational','Erotica','Fantasy','Gag Humor','Girls Love','Gore','Gourmet','Harem',
'Hentai','High Stakes Game','Historical','Horror','Idols (Female)','Idols (Male)','Isekai','Iyashikei','Josei','Kids','Love Polygon',
'Magical Sex Shift','Mahou Shoujo','Martial Arts','Mecha','Medical','Military','Music','Mystery','Mythology','Organized Crime','Otaku Culture',
'Parody','Performing Arts','Pets','Psychological','Racing','Reincarnation','Reverse Harem','Romance','Romantic Subtext','Samurai','School','Sci-Fi',
'Seinen','Shoujo','Shounen','Showbiz','Slice of Life','Space','Sports','Strategy Game','Super Power','Supernatural','Survival','Suspense','Team Sports',
'Time Travel','Vampire','Video Game','Visual Arts','Workplace']

merged_df_1 = merged_df[all_col]

all_col_word = ['user_id', 'user_score', 'num_watched_episodes','created_at','updated_at', 'anime_id', 'score','members', 'type', 'studio', 'episode_count', 
'Action','Adult Cast','Adventure','Anthropomorphic','Avant Garde','Award Winning','Boys Love','CGDCT','Childcare','Combat Sports','Comedy',
'Crossdressing','Delinquents','Detective','Drama','Ecchi','Educational','Erotica','Fantasy','Gag Humor','Girls Love','Gore','Gourmet','Harem',
'Hentai','High Stakes Game','Historical','Horror','Idols (Female)','Idols (Male)','Isekai','Iyashikei','Josei','Kids','Love Polygon',
'Magical Sex Shift','Mahou Shoujo','Martial Arts','Mecha','Medical','Military','Music','Mystery','Mythology','Organized Crime','Otaku Culture',
'Parody','Performing Arts','Pets','Psychological','Racing','Reincarnation','Reverse Harem','Romance','Romantic Subtext','Samurai','School','Sci-Fi',
'Seinen','Shoujo','Shounen','Showbiz','Slice of Life','Space','Sports','Strategy Game','Super Power','Supernatural','Survival','Suspense','Team Sports',
'Time Travel','Vampire','Video Game','Visual Arts','Workplace']
for i in range(100):
    all_col_word.append(str(i))

merged_df_2 = merged_df[all_col_word]

merged_df_3 = merged_df.drop(['ranked', 'popularity', 'rating_count'], axis=1)

with open('./dataset/processed_dataset/merged/merged_df_1.pkl','wb') as f:
    pickle.dump(merged_df_1, f)

with open('./dataset/processed_dataset/merged/merged_df_2.pkl','wb') as f:
    pickle.dump(merged_df_2, f)

with open('./dataset/processed_dataset/merged/merged_df_3.pkl','wb') as f:
    pickle.dump(merged_df_3, f)
